[MISC] Move island.py under the constraint subpackage.
